Better random test support.
This commit has a few small related changes that, in the end, allow
tests to be run in a different random order each time they are
executed.

The random number that was chosen is provided in the summary output so
that the test can be re-run witht the same random number at a later
time.

The previous behavior is preserved.  The new behavior requires macro
and/or command line changes.